+++ src/app/pages/vc-selector/vc-selector.page.ts
@@ -46,17 +46,6 @@ export class VcSelectorPage implements OnInit {
     redirectUri: '',
   };
 
-  public closeButton = [
-    {
-      text: this.translate.instant('vc-selector.close'),
-      role: 'confirm',
-      handler: () => {
-        this.sendCredentialAlert = false;
-        this.router.navigate(['/tabs/home']);
-      },
-    },
-  ];
-
   public constructor(
     private router: Router,
     private walletService: WalletService,
@@ -111,7 +100,7 @@ export class VcSelectorPage implements OnInit {
       this._VCReply.selectedVcList = this.selCredList;
       this.walletService.executeVC(this._VCReply).subscribe({
         next: () => {
-          this.sendCredentialAlert = true;
+          this.okMessage();
         },
         error: async (err) => {
           console.error(err);
@@ -128,21 +117,41 @@ export class VcSelectorPage implements OnInit {
   }
   private async errorMessage(){
     const alert = await this.alertController.create({
-      header: this.translate.instant('vc-selector.ko-message'),
-      message: '<img src="../assets/icon/Tick/close-circle-outline.svg" color="red"alt="g-maps" class="vs-selector-alert">',
+      message: `
+        <div style="display: flex; align-items: center; gap: 50px;">
+          <ion-icon name="alert-circle-outline"></ion-icon>
+          <span>${this.translate.instant('vc-selector.ko-message')}</span>
+        </div>
+      `,
       buttons: [
         {
-          text: this.translate.instant('confirmation.ok'),
+          text: this.translate.instant('vc-selector.close'),
           role: 'ok',
         },
       ],
-      cssClass:"custom-close-button"
+      cssClass:'custom-alert-error'
     });
 
     await alert.present();
     await alert.onDidDismiss();
   }
-  public setOpen(isOpen: boolean) {
-    this.sendCredentialAlert = isOpen;
+
+  private async okMessage() {
+    const alert = await this.alertController.create({
+      message: `
+        <div style="display: flex; align-items: center; gap: 50px;">
+          <ion-icon name="checkmark-circle-outline" ></ion-icon>
+          <span>${this.translate.instant('vc-selector.ok-header')}</span>
+        </div>
+      `,
+      cssClass: 'custom-alert-ok',
+    });
+  
+    await alert.present();
+  
+    setTimeout(async () => {
+      await alert.dismiss();
+      this.router.navigate(['/tabs/home']);
+    }, 2000);
   }
 }